refactor notifier
